pyproject.toml
[project]
name = "yolo-object-detection-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless>=4.0.0",
    "pillow>=12.1.1",
    "streamlit>=1.55.0",
    "ultralytics>=8.4.22",
]
